requirements.txt
ultralytics
opencv-python-headless
matplotlib
requests
python-dotenv
sagemaker
